+++ include/skl.h
@@ -48,7 +48,7 @@
 #endif
 
 #if defined(__riscv_xsfmm32a32f)
-#include "gemm/xsfmm/gemm_a1b01_f32c_f32_f32_xsfmm32a32f.h"
+#include "gemm/xsfmm/gemm_f32c_f32_f32_xsfmm32a32f.h"
 #endif
 
 #if defined(__riscv_xsfvqdotq)
